pyproject.toml
[tool.poetry]
name = "vnd-eng-practices-course"
version = "0.1.0"
description = ""
authors = ["Vladislav Nosivskoy <vladnosiv@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.23.5"
scikit-learn = "^1.1.3"
neptune-client = "^0.16.13"

[tool.poetry.group.dev.dependencies]
python = "^3.9"
numpy = "^1.23.5"
scikit-learn = "^1.1.3"
pytest = "^7.2.0"
black = "^22.10.0"
isort = "^5.10.1"
flake8 = "^6.0.0"
flake8-annotations = "^2.9.1"
flake8-pytest = "^1.4"
flake8-import-order = "^0.18.2"
pep8-naming = "^0.13.2"
flake8-commas = "^2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
